+++ src/imio/smartweb/core/contents/sections/selections/view.py
@@ -3,6 +3,10 @@
 from imio.smartweb.core.contents.sections.views import CarouselOrTableSectionView
 from imio.smartweb.core.utils import batch_results
 
+import logging
+
+logger = logging.getLogger("imio.smartweb.core")
+
 
 class SelectionsView(CarouselOrTableSectionView):
     """Selection Section view"""
@@ -12,6 +16,12 @@ class SelectionsView(CarouselOrTableSectionView):
         items = [rel.to_object for rel in self.context.selected_items]
         results = []
         for item in items:
+            if item is None:
+                # TODO we need to understand why relations break and fix that
+                logger.warning(
+                    f"Broken selection item in {self.context.absolute_url()}"
+                )
+                continue
             url = item.absolute_url()
             description = item.description or ""
             has_image = True if getattr(item.aq_base, "image", None) else False